// Frontend/src/app/Shared/Models/reserva-diamodel.ts
export class ReservaDia{
    fecha?: Date;
    sectorId?: number;
    ventaId?: number;
    sectorName?: string;
    descripcion?: any;
    clienteId?: number;
    clienteName?: string;
    relevanceValue?: number;
    productParentId?: number;
    productParentName?: any;
    productId?: number;
    productName?: string;
    cantidadVenta?: number;
    cantidadVentaEstandar?: number;
    cantidadCompra?: number;
    cantidadCompraEstandar?: number;
    unidadMedida?: string;
    precioIngresadoVenta?: number;
    amortizacion?: number;
    stock?: number;
    stockSobrante?: number;
    stockInicial?: number;
    observacion?:string;
  }
  